FEA: Rename turbine_metadata to turbine to avoid redundant description
